fix(ui): keep sidebar files within line budget
